Move files to right directory
