+++ src/api/authService.js
@@ -11,33 +11,5 @@ const authService = {
       throw handleError(error);
     }
   },
-
-  signup: async (userData) => {
-    try {
-      const response = await axiosInstance.post("/auth/signup", userData);
-      return response.data;
-    } catch (error) {
-      throw handleError(error);
-    }
-  },
-
-  logout: async () => {
-    try {
-      const response = await axiosInstance.post("/auth/logout");
-      return response.data;
-    } catch (error) {
-      throw handleError(error);
-    }
-  },
-
-  getUsername: async (user_email) => {
-    try {
-      const response = await axiosInstance.get(`/users/${user_email}`);
-      return response.data;
-    } catch (error) {
-      throw handleError(error);
-    }
-  },
 };
-
 export default authService;
