+++ src/components/Footer.jsx
@@ -9,7 +9,7 @@ import {
   IconButton,
 } from "@mui/material";
 import { Facebook, Twitter, Instagram, LinkedIn } from "@mui/icons-material";
-import colors from "../utils/colors";
+import colors from "../constants/colors";
 
 const Footer = () => {
   return (
@@ -42,12 +42,20 @@ const Footer = () => {
             </Typography>
             <Grid container spacing={2}>
               <Grid item xs={6}>
-                <Link href="#faq-home-main" color="secondary.contrastText" underline="none">
+                <Link
+                  href="#faq-home-main"
+                  color="secondary.contrastText"
+                  underline="none"
+                >
                   FAQ
                 </Link>
               </Grid>
               <Grid item xs={6}>
-                <Link href="#contact-us-home-main" color="secondary.contrastText" underline="none">
+                <Link
+                  href="#contact-us-home-main"
+                  color="secondary.contrastText"
+                  underline="none"
+                >
                   Contact Us
                 </Link>
               </Grid>
