+++ src/pages/AlumniDirectory.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Container, Box, Typography, Toolbar, Grid2 } from "@mui/material";
-import { alumniData as alumniFakeData, another } from "../utils/alumnidata";
+import { alumniData as alumniFakeData, another } from "../constants/alumnidata";
 import AlumniCard from "../components/AlumniCard";
-import { getUsers } from "../features/alumniUsersSlice";
+import { getUsers } from "../features/alumniSlice";
 
 const alumniData_1 = new Array(5).fill(alumniFakeData);
 const alumniData = alumniData_1.concat(new Array(5).fill(another));
