+++ src/store/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import authReducer from "../features/authSlice";
-import alumniUsersReducer from "../features/alumniUsersSlice";
+import alumniUsersReducer from "../features/alumniSlice";
 import announcementsReducer from "../features/announcementsSlice";
 
 const store = configureStore({
